Refactor category-related components to enhance loading state management
- Integrated `Suspense` and `withFallback` into `CategoryInline`, `CategoryValue`, `ListingCard`, and `BuyerInfo` components for improved loading experiences.
- Removed redundant `Pending` and `Data` components to streamline the structure and enhance maintainability.
- Updated various components to utilize colocated fallback UIs, ensuring consistent loading states across the application.
- Enhanced user interaction by adding `data-action` attributes for better accessibility and feedback.
