+++ apps/app/src/app/@buyer/listing/SellerInfo/SellerInfo.tsx
@@ -1,10 +1,17 @@
-import { type FC, Suspense } from "react";
-import { Data } from "./Data";
-import { Pending } from "./Pending";
+import type { MarkSuspense } from "@use-pico/client/type";
+import { Container, SpinnerContainer } from "@use-pico/client/ui/container";
+import { Group } from "@use-pico/client/ui/group";
+import { Status } from "@use-pico/client/ui/status";
+import { translator } from "@use-pico/common/translator";
+import { withFallback } from "@use-pico/client/utils";
+import { withListingSellerInfoQuery } from "@zbav-se.me/sdk/query/buyer/listing";
+import { SearchIcon } from "@zbav-se.me/ui/icon";
+import { Header } from "./Header";
+import { Score } from "./Score";
 
 export namespace SellerInfo {
-	export interface Props extends Omit<Data.Props, "_suspense"> {
-		//
+	export interface Props extends Container.Props, MarkSuspense.Props {
+		listingId: string;
 	}
 }
 
@@ -14,13 +21,41 @@ export namespace SellerInfo {
  *
  * @see apps/app/src/app//listing/ui/SellerInfoButton.tsx
  */
-export const SellerInfo: FC<SellerInfo.Props> = (props) => {
+export const SellerInfo = withFallback(({ _suspense, listingId, ui, ...props }: SellerInfo.Props) => {
+	const { data: sellerInfo } = withListingSellerInfoQuery.useSuspenseQuery({
+		listingId,
+	});
+
 	return (
-		<Suspense fallback={<Pending />}>
-			<Data
-				_suspense={"I know"}
-				{...props}
-			/>
-		</Suspense>
+		<Container
+			ui={{
+				flow: "vertical",
+				gap: "default",
+				height: "full",
+				...ui,
+			}}
+			{...props}
+		>
+			<Header sellerInfo={sellerInfo} />
+
+			{sellerInfo.events ? (
+				<Group>
+					<Score rank={sellerInfo.events.score.rank} />
+				</Group>
+			) : (
+				<Status
+					icon={SearchIcon}
+					textTitle={translator.text("Listing seller info not available (title)")}
+					textMessage={translator.text("Listing seller info not available (message)")}
+					ui={{
+						tone: "brand",
+						theme: "light",
+						inner: "2xl",
+						opacity: "6",
+					}}
+					className={"text-center"}
+				/>
+			)}
+		</Container>
 	);
-};
+}, SpinnerContainer);
