+++ apps/app/src/app/@seller/transaction/ui/BuyerInfo/BuyerInfo.tsx
@@ -1,10 +1,18 @@
-import { type FC, Suspense } from "react";
-import { Data } from "./Data";
-import { Pending } from "./Pending";
+import { useLocale } from "@use-pico/client/hook";
+import type { MarkSuspense } from "@use-pico/client/type";
+import { Container, LabelValue, SpinnerContainer } from "@use-pico/client/ui/container";
+import { Status } from "@use-pico/client/ui/status";
+import { toTimeDiff } from "@use-pico/common/time";
+import { translator } from "@use-pico/common/translator";
+import { withFallback } from "@use-pico/client/utils";
+import { withTransactionBuyerInfoQuery } from "@zbav-se.me/sdk/query/seller/transaction";
+import { SearchIcon } from "@zbav-se.me/ui/icon";
+import { Events } from "./Events";
+import { Score } from "./Score";
 
 export namespace BuyerInfo {
-	export interface Props extends Omit<Data.Props, "_suspense"> {
-		//
+	export interface Props extends Container.Props, MarkSuspense.Props {
+		transactionId: string;
 	}
 }
 
@@ -14,13 +22,52 @@ export namespace BuyerInfo {
  *
  * @see apps/app/src/app//transaction/ui/BuyerInfoButton.tsx
  */
-export const BuyerInfo: FC<BuyerInfo.Props> = (props) => {
+export const BuyerInfo = withFallback(({ _suspense, transactionId, ui, ...props }: BuyerInfo.Props) => {
+	const locale = useLocale();
+	const { data } = withTransactionBuyerInfoQuery.useSuspenseQuery({
+		where: {
+			id: transactionId,
+		},
+	});
+	const events = data.events;
+
 	return (
-		<Suspense fallback={<Pending />}>
-			<Data
-				_suspense={"I know"}
-				{...props}
+		<Container
+			ui={{
+				flow: "vertical",
+				gap: "default",
+				...ui,
+			}}
+			{...props}
+		>
+			<LabelValue
+				textLabel={translator.text("User registered (label)")}
+				textValue={toTimeDiff({
+					locale,
+					time: data.registered,
+					type: "relative",
+				})}
 			/>
-		</Suspense>
+
+			{events ? (
+				<>
+					<Events events={events} />
+					<Score rank={events.score.rank} />
+				</>
+			) : (
+				<Status
+					icon={SearchIcon}
+					textTitle={translator.text("Transaction buyer info not available (title)")}
+					textMessage={translator.text("Transaction buyer info not available (message)")}
+					ui={{
+						tone: "brand",
+						theme: "light",
+						inner: "2xl",
+						opacity: "6",
+					}}
+					className={"text-center"}
+				/>
+			)}
+		</Container>
 	);
-};
+}, SpinnerContainer);
