+++ apps/app/src/app/@seller/transaction/ui/button/BuyerInfoButton.tsx
@@ -4,7 +4,7 @@ import { Tx } from "@use-pico/client/ui/tx";
 import { translator } from "@use-pico/common/translator";
 import { CloseButton } from "@zbav-se.me/ui/button";
 import { BuyerIcon } from "@zbav-se.me/ui/icon";
-import { type FC, useState } from "react";
+import { Suspense, type FC, useState } from "react";
 import { BuyerInfo } from "../BuyerInfo/BuyerInfo";
 
 export namespace BuyerInfoButton {
@@ -46,12 +46,15 @@ export const BuyerInfoButton: FC<BuyerInfoButton.Props> = ({ transactionId, ...p
 					right: <CloseButton onClick={close} />,
 				})}
 			>
-				<BuyerInfo
-					transactionId={transactionId}
-					ui={{
-						inner: "default",
-					}}
-				/>
+				<Suspense fallback={<BuyerInfo.Fallback />}>
+					<BuyerInfo
+						_suspense={"I know"}
+						transactionId={transactionId}
+						ui={{
+							inner: "default",
+						}}
+					/>
+				</Suspense>
 			</BottomSheet>
 		</>
 	);
