--- apps/app/src/app/@session/category/ui/CategoryInline/Pending.tsx
@@ -1,17 +1,0 @@
-import { SpinnerContainer } from "@use-pico/client/ui/container";
-import type { FC } from "react";
-
-export namespace Pending {
-	export interface Props extends SpinnerContainer.Props {
-		//
-	}
-}
-
-export const Pending: FC<Pending.Props> = (props) => {
-	return (
-		<SpinnerContainer
-			data-ui={"CategoryValueListContent-[SpinnerContainer.category-inline]"}
-			{...props}
-		/>
-	);
-};
