+++ apps/app/src/app/@session/category/ui/CategorySelect/CategoryItem/index.ts
@@ -1,0 +1,1 @@
+export { CategoryItem } from "./CategoryItem";
