+++ apps/app/src/app/@session/category/ui/CategorySelect/ListContainer/index.ts
@@ -1,0 +1,1 @@
+export { ListContainer } from "./ListContainer";
